Development (#199)

+++ immuneML/encodings/protein_embedding/ESMCEncoder.py
@@ -1,13 +1,14 @@
-import numpy as np
 import logging
+
+import numpy as np
+
 from immuneML.data_model.AIRRSequenceSet import AIRRSequenceSet
 from immuneML.data_model.SequenceParams import RegionType
 from immuneML.data_model.datasets.Dataset import Dataset
-from immuneML.encodings.EncoderParams import EncoderParams
 from immuneML.encodings.protein_embedding.ProteinEmbeddingEncoder import ProteinEmbeddingEncoder
+from immuneML.util.Logger import log_memory_usage
 from immuneML.util.NumpyHelper import NumpyHelper
 from immuneML.util.ParameterValidator import ParameterValidator
-from immuneML.util.Logger import log_memory_usage
 
 
 class ESMCEncoder(ProteinEmbeddingEncoder):
@@ -121,7 +122,7 @@ class ESMCEncoder(ProteinEmbeddingEncoder):
                     sequence_id=sequence_id
                 )
 
-            batch_embeddings = output.embeddings.cpu().numpy().mean(axis=1)
+            batch_embeddings = output.embeddings.cpu().float().numpy().mean(axis=1)
             embeddings[i:batch_end] = batch_embeddings
 
             del output, tokens, sequence_id
